First draft typer-based CLI
